Move 'clickable' to private folder
